logging: fix old methods on BlasMod class
